feat: add function createMyServer
