+++ tsconfig.json
@@ -13,6 +13,6 @@
   "ts-node": {
     "esm": true
   },
-  "include": ["src/*.ts", "src/index.ts"],
+  "include": ["src/*.ts", "src/server.ts"],
   "exclude": ["node_modules", "dist"]
 }
